moves conditional logic tests to standalone spec and removes use client directive from custom server field
