+++ test/fields/e2e.spec.ts
@@ -494,87 +494,6 @@ describe('fields', () => {
     })
   })
 
-  describe('conditional logic', () => {
-    let url: AdminUrlUtil
-    beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, 'conditional-logic')
-    })
-
-    test('should toggle conditional field when data changes', async () => {
-      await page.goto(url.create)
-      const toggleField = page.locator('label[for=field-toggleField]')
-
-      // eslint-disable-next-line playwright/no-conditional-in-test
-      if (!(await toggleField.isChecked())) {
-        await toggleField.click()
-      }
-
-      await expect(page.locator('label[for=field-toggleField]')).toBeChecked()
-      const fieldWithCondition = page.locator('input#field-fieldWithCondition')
-      await expect(fieldWithCondition).toBeVisible()
-    })
-
-    test('should toggle conditional custom client field', async () => {
-      await page.goto(url.create)
-      const toggleField = page.locator('label[for=field-toggleField]')
-
-      // eslint-disable-next-line playwright/no-conditional-in-test
-      if (!(await toggleField.isChecked())) {
-        await toggleField.click()
-      }
-
-      await expect(page.locator('label[for=field-toggleField]')).toBeChecked()
-      const customFieldWithCondition = page.locator('#custom-client-field')
-      await expect(customFieldWithCondition).toBeVisible()
-    })
-
-    test('should toggle conditional custom server field', async () => {
-      await page.goto(url.create)
-      const toggleField = page.locator('label[for=field-toggleField]')
-
-      // eslint-disable-next-line playwright/no-conditional-in-test
-      if (!(await toggleField.isChecked())) {
-        await toggleField.click()
-      }
-
-      await expect(page.locator('label[for=field-toggleField]')).toBeChecked()
-      const customFieldWithCondition = page.locator('#custom-server-field')
-      await expect(customFieldWithCondition).toBeVisible()
-    })
-
-    test('should show conditional field based on user data', async () => {
-      await page.goto(url.create)
-      const userConditional = page.locator('input#field-userConditional')
-      await expect(userConditional).toBeVisible()
-    })
-
-    test('should show conditional field based on fields nested within data', async () => {
-      await page.goto(url.create)
-
-      const parentGroupFields = page.locator(
-        'div#field-parentGroup > .group-field__wrap > .render-fields',
-      )
-      await expect(parentGroupFields).toHaveCount(1)
-
-      const toggle = page.locator('label[for=field-parentGroup__enableParentGroupFields]')
-      await toggle.click()
-
-      const toggledField = page.locator('input#field-parentGroup__siblingField')
-
-      await expect(toggledField).toBeVisible()
-    })
-
-    test('should show conditional field based on fields nested within siblingData', async () => {
-      await page.goto(url.create)
-
-      const toggle = page.locator('label[for=field-parentGroup__enableParentGroupFields]')
-      await toggle.click()
-
-      const fieldRelyingOnSiblingData = page.locator('input#field-reliesOnParentGroup')
-      await expect(fieldRelyingOnSiblingData).toBeVisible()
-    })
-  })
-
   describe('tabs', () => {
     // tabsFieldsSlug is used for testing tabs
     let tabsFieldsUrl: AdminUrlUtil
